Fix images not displaying correctly in edit mode
