+++ src/customizations/volto/components/manage/Blocks/Image/View.jsx
@@ -3,14 +3,9 @@
  * @module components/manage/Blocks/Image/View
  */
 
-import { UniversalLink } from '@plone/volto/components';
-import {
-  flattenToAppURL,
-  isInternalURL,
-  withBlockExtensions,
-} from '@plone/volto/helpers';
-import cx from 'classnames';
+import { withBlockExtensions } from '@plone/volto/helpers';
 import PropTypes from 'prop-types';
+import { Body } from './Body';
 
 /**
  * View image block class.
@@ -19,69 +14,7 @@ import PropTypes from 'prop-types';
  */
 export const View = ({ data }) => {
   const href = data?.href?.[0]?.['@id'] || '';
-  return (
-    <figure
-      className={cx('nsw-media', {
-        [`nsw-media--${data.captionBackgroundColour}`]: !!data.captionBackgroundColour,
-        [`nsw-media--${data.size}`]:
-          !!data.size &&
-          !['left', 'right'].includes(data.align) &&
-          data.size !== 'fullWidth',
-        [`nsw-media--${data.align}-${data.size}`]:
-          !!data.size && !!data.align && data.size !== 'fullWidth',
-      })}
-    >
-      {data.url && (
-        <>
-          {(() => {
-            const image = (
-              <img
-                className={cx({
-                  'full-width': data.align === 'full',
-                  large: data.size === 'l',
-                  medium: data.size === 'm',
-                  small: data.size === 's',
-                })}
-                src={
-                  isInternalURL(data.url)
-                    ? // Backwards compat in the case that the block is storing the full server URL
-                      (() => {
-                        if (data.size === 'l')
-                          return `${flattenToAppURL(data.url)}/@@images/image`;
-                        if (data.size === 'm')
-                          return `${flattenToAppURL(
-                            data.url,
-                          )}/@@images/image/preview`;
-                        if (data.size === 's')
-                          return `${flattenToAppURL(
-                            data.url,
-                          )}/@@images/image/mini`;
-                        return `${flattenToAppURL(data.url)}/@@images/image`;
-                      })()
-                    : data.url
-                }
-                alt={data.alt || ''}
-                loading="lazy"
-              />
-            );
-            if (href) {
-              return (
-                <UniversalLink
-                  href={href}
-                  openLinkInNewTab={data.openLinkInNewTab}
-                >
-                  {image}
-                </UniversalLink>
-              );
-            } else {
-              return image;
-            }
-          })()}
-        </>
-      )}
-      {data.caption ? <figcaption>{data.caption}</figcaption> : null}
-    </figure>
-  );
+  return <Body data={data} href={href} />;
 };
 
 /**
